precharge fixed since last time so that it starts true and always fails to true. Also fixed receiving battery messages

+++ src/precharge.h
@@ -10,7 +10,7 @@
 
 typedef struct {
 	//inputs
-	uint16_t voltage;
+	float voltage;
     bool relay_state;
 	
 	uint16_t pack_dlc;	//Max current the battery can output
